Create command infrastructure and use it to save settings.
Commands are indicated by a POST request called "command" which must refer to a
method of the jmct_Admin class. This method must have a "@command" doc comment
to verify that it is indeed a command.

This infrastructure is a security measure to validate the "command" parameter
and ensure that no jiggery-pokery is taking place.